+++ CHANGES.md
@@ -7,10 +7,52 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 Use this section to keep track of changes in the works.
 
+### Added
+* bit combo hierarchy in `ClusterSubspace` for fitting hierarchy constraints.
+  [\#106](https://github.com/CederGroupHub/smol/pull/106)
+  ([qchempku2017](https://github.com/qchempku2017))
+* data indices in `StructureWrangler` to keep track of training/test splits,
+  duplicate sets, etc.
+  [\#108](https://github.com/CederGroupHub/smol/pull/108)
+  ([lbluque](https://github.com/lbluque))
+* `ClusterSubspace.cutoffs` property to obtain tight cutoffs of included
+   orbits.
+   [\#108](https://github.com/CederGroupHub/smol/pull/108)
+   ([lbluque](https://github.com/lbluque))
+* Added properties to get orbit and ordering multiplicities of corr functions.
+[\#102](https://github.com/CederGroupHub/smol/pull/102)
+([lbluque](https://github.com/lbluque))
+  - `function_ordering_multiplicities`, `function_total_multiplicities`
+* Added helpful methods/properties to `ClusterSubspace` to get corr function
+  indices based on cluster diameter cutoffs and/or cluster sizes.
+[\#102](https://github.com/CederGroupHub/smol/pull/102)
+([lbluque](https://github.com/lbluque))
+  - `orbits_by_cutoffs`, `function_inds_by_cutoffs`, `function_inds_by_size`.
+
 ### Changed
+* filter functions in `smol.wrangling` replaced with functions returning
+  indices corresponding to structures to keep. This can be used saving indices
+  with `StructureWrangler.add_data_indices`.
+[\#102](https://github.com/CederGroupHub/smol/pull/108)
+([lbluque](https://github.com/lbluque))
 * Cleanup of sites, active sites and restricted sites in `Sublattice`
 [\#95](https://github.com/CederGroupHub/smol/pull/95)
   ([juliayang](https://github.com/juliayang))
+* Make feature matrix optional when creating a `ClusterExpansion` construction.
+[\#102](https://github.com/CederGroupHub/smol/pull/102)
+([lbluque](https://github.com/lbluque))
+* Renamed `ncorr_functions_per_orbit` -> `num_functions_per_orbit` in
+  `ClusterSubspace` and `ClusterExpansion.convert_eci` ->
+  `ClusterExpansion.convert_coefs`
+[\#102](https://github.com/CederGroupHub/smol/pull/102)
+([lbluque](https://github.com/lbluque))
+* Changed orthonormalization of site basis to use `np.linalg.qr`.
+[\#102](https://github.com/CederGroupHub/smol/pull/102)
+([lbluque](https://github.com/lbluque))
+* Changed cython corr functions to reduce Python interaction in loop
+(~1.5x faster)
+[\#102](https://github.com/CederGroupHub/smol/pull/102)
+([lbluque](https://github.com/lbluque))
  
 ### Fixed
 * Fixed `Sublattice` serialization, saving/loading `SiteSpaces`.